requirements-dev.txt
black=19.10b0
pre-commit
coveralls
filelock
flake8
pillow<7
imagehash
pytest
pytest-cov
pytest-xdist
sphinx
